move component to be forum-general
